Use onpopstate event to rerender on back button
The onpopstate event triggers as soon as the back button is triggered,
rather than waiting up to 100ms for it to happen. An additional bonus of
not polling for the URL change is that the app won't wake up the CPU if
the user isn't using the app. This will help keep from draining
batteries in laptops and mobile devices.

This commit also fixes a weird bug where Application#watch_url was
accidentally defined within Application#trap_clicks. It only worked as
intended because trap_clicks was called before watch_url.