Only render when the app is visible
While in a background tab or while the browser is hidden behind another
window or minimized, rerendering can be wasteful. If we rerender N
times, the user will not see N - 1 of them, so we want to wait to
rerender when the app becomes visible again.

This may reduce CPU usage within the app and extend the life of
battery-powered devices.