Improve render coalescing
Prior to this commit, calling Application#render twice in a row would
actually trigger two renders; the first render was always eager. It
turns out that the scheduled render was quick enough that it didn't
matter if we waited to do the first one.

This also resolves a weird bug in Safari where the app would render
twice if the web inspector was open. Turns out, when the web inspector
is open, Safari triggers a popstate event after page load. This wasn't
intended to fix that bug because it really only happens in development,
but it's a pretty cool side effect.