Whitelist vs blacklist
Previously we were whitelisting values that attributes could be if we
wanted to use them as content instead (basically, attributes were
omitted and content was supplied), but this didn't allow for VirtualDOM
nodes. This commit changes that so that the attributes are only treated
as attributes if they are blank or a Hash.