Use requestAnimationFrame to schedule render
Previously, we were kinda rolling our own. This does the same thing but
is simpler. Calling app.render doesn't render right away, but instead
schedules a render to take place soon. Prior to this commit, it was
after a 1/60th-second timeout, which is likely to skip a frame in
between. Using requestAnimationFrame isn't supposed to do that, AFAIK.

We did still have to implement our own render throttling because
spamming requestAnimationFrame with the same function will happily
invoke every single one of those calls on the next animation frame,
which is what this entire construct is intended to protect from.