Inline StringUtils.camlize to JS
The String#gsub method gets wrapped in a try/catch block by Opal, which
the VM can't optimize. Using the direct JS version also removes a lot of
the flexibility of the gsub method that we don't need, which is good
since it comes with a performance penalty, as well.