Remove touch handlers from Link
These handlers were originally there to get rid of the gesture delay
when tapping links in Webkit. It was the only browser engine that didn't
yet optimize this when the app was unscaled. However, it is now in
Webkit and in the iOS Safari 9.1 beta, so we don't really need this
anymore.