Two small refactors to routing code
This commit includes two small refactors to two different parts of the router code. There was some inefficient processing of an array, so instead of a call to `.map` and then `.flatten!`, I've changed it to `.flat_map`, which is more performant.

Also, in `opal/clearwater/router.rb` there was some duplicated code around generating the parts of a path from a given path string, which I've extracted into a single method.