Hydrate existing server-rendered app
Previously, server-rendering a Clearwater app involved generating the
virtual DOM nodes and *replacing* what was rendered by the server. This
is pretty inefficient, though, because it has to throw away all of the
existing server-rendered nodes. This can be prohibitively expensive on
slower devices.

This commit lets Clearwater handle the existing DOM nodes as if they
were artifacts of a previous Clearwater render.