Remove node when unmounting
This removes a condition where, in a persistent component that holds
onto a DOMRef that proxies a video node, that video could stay in memory
for the life of the app. Even if it gets replaced on the next render, it
would simply be recycling memory but never truly freeing it.